Add trainer dashboard UI with some mock data
