+++ client/src/components/Dashboard/index.ts
@@ -1,0 +1,8 @@
+export { CheckInsCard } from './CheckInsCard';
+export { ClientsMissingPlanCard } from './ClientsMissingPlanCard';
+export { DashboardHeader } from './DashboardHeader';
+export { DashboardStats } from './DashboardStats';
+export { FocusTodayCard } from './FocusTodayCard';
+export { QuickActionsCard } from './QuickActionsCard';
+export { RecentClientsCard } from './RecentClientsCard';
+export { TodaysAppointmentsCard } from './TodaysAppointmentsCard';
